Move routes around, catch-all on [...filter]
