+++ src/routes/(app)/(public)/[...type]/+page.svelte
